extract components from Pagination
